Cap emulation to 60.0 fps
